proyecto v3

+++ public/js/singleCart.js
@@ -1,8 +1,7 @@
 const productCards = document.querySelectorAll('.product-item');
 const cId = document.querySelector('#cart-id').textContent.slice(8);
 const btnVaciar = document.querySelector('#btn-vaciar');
-console.log(btnVaciar)
-console.log(productCards)
+
 productCards.forEach(card => {
   const pId = card.querySelector('#product-id').textContent;
   const btnEliminar = card.querySelector('#btn-eliminar');
@@ -27,17 +26,29 @@ productCards.forEach(card => {
 });
 
 btnVaciar.addEventListener('click', async() => {
-  try{
-    const response = await fetch(`/api/carts/${cId}/products`, {
-      method: 'DELETE'
+  Swal.fire({
+      title: '¿Estás seguro de vaciar el carrito?',
+      icon: 'warning',
+      showCancelButton: true,
+      confirmButtonColor: '#d33',
+      cancelButtonColor: '#3085d6',
+      confirmButtonText: 'Vaciar',
+      cancelButtonText: 'Cancelar'
+    }).then(async (result) => {
+      if(result.isConfirmed){
+        try{
+          const response = await fetch(`/api/carts/${cId}/products`, {
+            method: 'DELETE'
+          });
+          if(response.ok){
+            console.log(`Carrito ${cId} vaciado`);
+            window.location.reload();
+          } else{
+            console.log('Error al vaciar el carrito');
+          }
+        } catch(error){
+          console.log('Error al vaciar el carrito:', error);
+        }
+      }
     });
-    if(response.ok){
-      console.log(`Carrito ${cId} vaciado`);
-      window.location.reload();
-    } else{
-      console.log('Error al vaciar el carrito');
-    }
-  } catch(error){
-    console.log('Error al vaciar el carrito:', error);
-  }
 });
